requirements.txt
opencv-python>=4.8
mediapipe>=0.10
numpy>=1.25
pillow>=10.0
